+++ shop/cart/cart_modifiers_base.py
@@ -10,7 +10,7 @@ class BaseCartModifier(object):
     settings.SHOP_CART_MODIFIERS setting.
     """
         
-    def process_cart_item(self, cart_item):
+    def process_cart_item(self, cart_item, cart):
         """
         This will be called for every line item in the Cart:
         Line items typically contain: product, unit_price, quantity...
@@ -21,8 +21,11 @@ class BaseCartModifier(object):
         
         Overrides of this method should however update cart_item.current_total, 
         since it will potentially be used by other modifiers.
+        
+        The cart parameter is only used to let implementations access the current 
+        cart object (to store information for example)
         """
-        field = self.get_extra_cart_item_price_field(cart_item)
+        field = self.get_extra_cart_item_price_field(cart_item, cart)
         if field != None:
             price = field[1]
             cart_item.current_total = cart_item.current_total + price
@@ -48,7 +51,7 @@ class BaseCartModifier(object):
             cart.extra_price_fields.append(field)
         return cart
     
-    def get_extra_cart_item_price_field(self, cart_item):
+    def get_extra_cart_item_price_field(self, cart_item, cart):
         """
         Get an extra price field tuple for the current cart_item:
         
@@ -68,6 +71,9 @@ class BaseCartModifier(object):
         >>> return ('rebate', Decimal(-9))
 
         More examples can be found in shop.cart.modifiers.*
+        
+        The cart parameter is only providing a reference to the current cart for
+        extensibility and flexibility purposes.
         """
         return None # Does nothing by default