+++ shop/models/defaults/bases.py
@@ -185,7 +185,7 @@ class BaseCart(models.Model):
         self.subtotal_price = Decimal('0.0') # Reset the subtotal
 
         for item in items: # For each OrderItem (order line)...
-            self.subtotal_price = self.subtotal_price + item.update()
+            self.subtotal_price = self.subtotal_price + item.update(self)
             item.save()
         
         self.current_total = self.subtotal_price
@@ -239,7 +239,7 @@ class BaseCartItem(models.Model):
         self.line_total = Decimal('0.0')
         self.current_total = Decimal('0.0') # Used by cart modifiers
 
-    def update(self):
+    def update(self, cart):
         self.extra_price_fields = [] # Reset the price fields
         self.line_subtotal = self.product.get_price() * self.quantity
         self.current_total = self.line_subtotal
@@ -247,7 +247,7 @@ class BaseCartItem(models.Model):
         for modifier in cart_modifiers_pool.get_modifiers_list():
             # We now loop over every registered price modifier,
             # most of them will simply add a field to extra_payment_fields
-            modifier.process_cart_item(self)
+            modifier.process_cart_item(self, cart)
         
         self.line_total = self.current_total
         return self.line_total