+++ shop/models/defaults/managers.py
@@ -102,7 +102,7 @@ class OrderManager(models.Manager):
         # There, now move on to the order items.
         cart_items = CartItem.objects.filter(cart=cart)
         for item in cart_items:
-            item.update()
+            item.update(cart)
             i = OrderItem()
             i.order = o
             i.product_reference = item.product.id